Branch gabriel

+++ API_Banco/src/main/java/com/Gabriel/API_Banco/repository/Repositorio.java
@@ -3,5 +3,6 @@ package com.Gabriel.API_Banco.repository;
 import com.Gabriel.API_Banco.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Repositorio extends JpaRepository<Usuario, Integer> {
+public interface Repositorio extends JpaRepository<Usuario, Long> {
+
 }
